requirements.txt
certifi==2020.6.20
chardet==3.0.4
Django==2.2.12
django-appconf==1.0.4
django-ckeditor==5.9.0
django-colorfield==0.3.2
django-extensions==2.2.8
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.4.0
idna==2.10
importlib-metadata==1.7.0
pilkit==2.0
Pillow==7.2.0
PyJWT==1.7.1
python-dotenv==0.12.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.1.0
